Manually fix remaining linting errors
We should avoid making the Gemfile conditional on environment
variables. This can always be done locally, on the relatively
few occassions it's actually needed.

Note that the User model is not a normal ActiveRecord model and
instead defines its own 'update_attribute(s)' methods to make
it compatible with GDS SSO.